refactor: Annotation 기반으로 변경
