+++ app/src/main/java/com/techcourse/controller/RegisterViewController.java
@@ -2,12 +2,15 @@ package com.techcourse.controller;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.controller.asis.Controller;
+import nextstep.web.annotation.Controller;
+import nextstep.web.annotation.RequestMapping;
+import nextstep.web.support.RequestMethod;
 
-public class RegisterViewController implements Controller {
+@Controller
+public class RegisterViewController {
 
-    @Override
-    public String execute(HttpServletRequest req, HttpServletResponse res) throws Exception {
+    @RequestMapping(value = "/register/view", method = RequestMethod.GET)
+    public String execute(HttpServletRequest req, HttpServletResponse res) {
         return "/register.jsp";
     }
 }
